Refactored additional support for update journey
Post, get, validate - all working

tests passing for update journey
